services/ingestion/requirements.txt
paho-mqtt==1.6.1
influxdb-client==1.41.0
pydantic==1.10.17
python-dotenv==1.0.1
